parser-json-gcc: separate module for GccTreeDecoder
